+++ NEWS.md
@@ -3,6 +3,9 @@
 * htmlwidgets (e.g., leaflet) can now be used in Rd `@examples`. See
   `test-widgets` for examples.
 
+* `build_reference_index()`: Selectors that do not match topics now generate a warning.
+  If none of the specified selectors have a match, no topics are selected (#728).
+  
 * Support of multiple arguments in `\Sexpr{}` was fixed, eliminating `x must be a
   string or a R connection` errors when using `\doi` Rd tags (#738).